requirements/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
psycopg2-binary==2.9.4
django_extensions==3.2.1
pyOpenSSL==22.1.0
requests==2.28.1
flake8==5.0.4
black[jupyter]==22.10.0